+++ src/main/webapp/query-builder/filter-inputs/number-filter.tsx
@@ -6,7 +6,7 @@ import Box from '@material-ui/core/Box'
 import LinearProgress from '@material-ui/core/LinearProgress'
 import AttributeDropdown from '../filter/attribute-dropdown'
 import ComparatorDropdown from '../filter/comparator-dropdown'
-import sampleAttributeDefinitions from '../filter/sample-attribute-definitions'
+import sampleAttributeDefinitions from '../../sample-data/sample-attribute-definitions.json'
 import useAttributeDefinitions from '../../react-hooks/use-attribute-definitions'
 import { AttributeDefinition } from '../types'
 import { useState } from 'react'
