+++ src/main/webapp/query-builder/filter/attribute-dropdown.tsx
@@ -5,7 +5,7 @@ import TextField from '@material-ui/core/TextField'
 import { QueryFilterProps } from './individual-filter'
 import { getDefaultValue } from './filter-utils'
 import { getIn } from 'immutable'
-import sampleAttributeDefinitions from './sample-attribute-definitions'
+import sampleAttributeDefinitions from '../../sample-data/sample-attribute-definitions.json'
 import { Comparators } from './comparator-dropdown'
 import LinearProgress from '@material-ui/core/LinearProgress'
 import { AttributeDefinition } from '../types'
