+++ src/main/webapp/query-builder/filter/comparator-dropdown.tsx
@@ -6,7 +6,7 @@ import Popover from '@material-ui/core/Popover'
 import MenuItem from '@material-ui/core/MenuItem'
 import DropDownIcon from '@material-ui/icons/ArrowDropDown'
 import { Map, getIn } from 'immutable'
-import sampleAttributeDefinitions from './sample-attribute-definitions'
+import sampleAttributeDefinitions from '../../sample-data/sample-attribute-definitions.json'
 import { QueryFilter, AttributeDefinition } from '../types'
 import LinearProgress from '@material-ui/core/LinearProgress'
 import useAttributeDefinitions from '../../react-hooks/use-attribute-definitions'
