+++ src/main/webapp/query-builder/filter/individual-filter.tsx
@@ -5,7 +5,7 @@ import DateFilter from '../filter-inputs/date-filter'
 import BooleanFilter from '../filter-inputs/boolean-filter'
 import NumberFilter from '../filter-inputs/number-filter'
 import { getIn } from 'immutable'
-import sampleAttributeDefinitions from './sample-attribute-definitions'
+import sampleAttributeDefinitions from '../../sample-data/sample-attribute-definitions.json'
 import { deserialize, serialize } from './filter-serialization'
 import { QueryFilter, AttributeDefinition } from '../types'
 import useAttributeDefinitions from '../../react-hooks/use-attribute-definitions'
