+++ src/main/webapp/sample-data/filterTree.json
