+++ src/main/webapp/sample-data/sample-attribute-definitions.json
@@ -1,0 +1,52 @@
+[
+  {
+    "type": "XML",
+    "enums": [],
+    "id": "metadata"
+  },
+  {
+    "type": "BINARY",
+    "enums": [],
+    "id": "thumbnail"
+  },
+  {
+    "type": "BOOLEAN",
+    "enums": [],
+    "id": "phonetics"
+  },
+  {
+    "type": "DATE",
+    "enums": [],
+    "id": "created"
+  },
+  {
+    "type": "DOUBLE",
+    "enums": [],
+    "id": "media.bit.rate"
+  },
+  {
+    "type": "INTEGER",
+    "enums": [],
+    "id": "media.width-pixels"
+  },
+  {
+    "type": "LONG",
+    "enums": [],
+    "id": "ext.population"
+  },
+  {
+    "type": "GEOMETRY",
+    "enums": [],
+    "id": "anyGeo"
+  },
+  {
+    "type": "GEOMETRY",
+    "enums": [],
+    "id": "location"
+  },
+  {
+    "type": "STRING",
+    "enums": ["cat", "dog", "catdog"],
+    "id": "topic.vocabulary"
+  }
+]
