+++ tsconfig.json
@@ -11,7 +11,8 @@
     "lib": ["es6", "dom", "es2017"],
     "target": "es5",
     "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true
+    "allowSyntheticDefaultImports": true,
+    "resolveJsonModule": true
   },
   "include": ["**/*.ts", "**/*.tsx"],
   "exclude": ["node_modules"]
